% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecAlternatives.R
\name{Spec}
\alias{Spec}
\alias{SpecFrequency0A}
\alias{SpecFrequency0B}
\alias{SpecFrequency0C}
\title{\code{spec} creation}
\usage{
Spec(...)

SpecFrequency0A(
  protectZeros = TRUE,
  extend0 = TRUE,
  primary = PrimaryDefault,
  candidates = CandidatesDefault,
  singleton = SingletonDefault,
  ...
)

SpecFrequency0B()

SpecFrequency0C()
}
\arguments{
\item{...}{Further arguments}

\item{protectZeros}{Suppression parameter.
When \code{TRUE}, cells with zero frequency or value are set as primary suppressed.
Using the default \code{primary} function, \code{protectZeros} is by default set to \code{TRUE}. See details.}

\item{extend0}{Data is automatically extended by \code{Extend0} when \code{TRUE}.
Can also be set to \code{"all"} which means that input codes in hierarchies are considered in addition to those in data.
Parameter \code{extend0} can also be specified as a list meaning parameter \code{varGroups} to \code{Extend0}.}

\item{primary}{GaussSuppression input or a function generating it (see details) Default: \code{\link{PrimaryDefault}}}

\item{candidates}{GaussSuppression input or a function generating it (see details) Default: \code{\link{CandidatesDefault}}}

\item{singleton}{GaussSuppression input or a function generating it (see details) Default: \code{\link{SingletonDefault}}}
}
\value{
A named list of arguments to be used as \code{spec}  input to \code{\link{GaussSuppressionFromData}}.
}
\description{
To make input to \code{\link{GaussSuppressionFromData}}

Function \code{Spec} is a simple wrapper for \code{\link{list}}.
Other functions are made for specific purposes.
}
\examples{
SpecFrequency0A()
SpecFrequency0B()
identical(SpecFrequency0A(), SpecFrequency0C())
}
