% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecAlternatives.R
\name{SuppressFrequency}
\alias{SuppressFrequency}
\alias{SuppressFrequency0A}
\alias{SuppressFrequency0B}
\alias{SuppressFrequency0_non_spec}
\title{SuppressFrequency}
\usage{
SuppressFrequency(
  data,
  dimVar = NULL,
  freqVar = NULL,
  maxN = 3,
  spec = SpecFrequency0A(),
  ...
)

SuppressFrequency0A(..., spec = SpecFrequency0A())

SuppressFrequency0B(..., spec = SpecFrequency0B())

SuppressFrequency0_non_spec(
  ...,
  protectZeros = TRUE,
  extend0 = TRUE,
  primary = PrimaryDefault,
  candidates = CandidatesDefault,
  singleton = SingletonDefault
)
}
\arguments{
\item{data}{Input data as a data frame}

\item{dimVar}{The main dimensional variables and additional aggregating variables. This parameter can be  useful when hierarchies and formula are unspecified.}

\item{freqVar}{A single variable holding counts (name or number).}

\item{maxN}{Suppression parameter. Cells with frequency \verb{<= maxN} are set as primary suppressed.
Using the default \code{primary} function, \code{maxN} is by default set to \code{3}. See details.}

\item{spec}{\code{NULL} or a named list of arguments that will act as default values.}

\item{...}{Further arguments}

\item{protectZeros}{Suppression parameter.
When \code{TRUE}, cells with zero frequency or value are set as primary suppressed.
Using the default \code{primary} function, \code{protectZeros} is by default set to \code{TRUE}. See details.}

\item{extend0}{Data is automatically extended by \code{Extend0} when \code{TRUE}.
Can also be set to \code{"all"} which means that input codes in hierarchies are considered in addition to those in data.
Parameter \code{extend0} can also be specified as a list meaning parameter \code{varGroups} to \code{Extend0}.}

\item{primary}{GaussSuppression input or a function generating it (see details) Default: \code{\link{PrimaryDefault}}}

\item{candidates}{GaussSuppression input or a function generating it (see details) Default: \code{\link{CandidatesDefault}}}

\item{singleton}{GaussSuppression input or a function generating it (see details) Default: \code{\link{SingletonDefault}}}
}
\description{
SuppressFrequency
}
\examples{
z1 <- SSBtoolsData("z1")
a1 <- GaussSuppressionFromData(data = z1, dimVar = 1:2, freqVar = 3, spec = SpecFrequency0A())
a2 <- SuppressFrequency(data = z1, dimVar = 1:2, freqVar = 3)
a3 <- SuppressFrequency0A(data = z1, dimVar = 1:2, freqVar = 3)
a4 <- SuppressFrequency0B(data = z1, dimVar = 1:2, freqVar = 3)
a5 <- SuppressFrequency0_non_spec(data = z1, dimVar = 1:2, freqVar = 3)
identical(a1, a2)
identical(a1, a3)
identical(a1, a4)
identical(a1, a5)
}
